Responsible-AI-Security-API/wrapper/requirements/requirement.txt
uvicorn==0.30.1
pyyaml==6.0.1
fastapi==0.111.0
watchfiles==0.22.0
requests==2.32.3
python-multipart==0.0.9
email-validator==2.2.0
adversarial-robustness-toolbox==1.16.0
openpyxl==3.1.4
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
scikit-learn==1.1.3
xgboost==2.1.0
tensorflow==2.15.0
keras==2.15.0
pymongo==4.7.3
python-dotenv==1.0.1
torch==2.3.1
pdfkit==1.0.0
reportlab==4.2.0
PyPDF2==3.0.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
mongomock==4.1.2
python-jose[cryptography]
pyjwt
lib/aicloudlibs-0.1.0-py3-none-any.whl
lib/adversarial_robustness_toolbox-1.16.0-py3-none-any.whl
#Un-Comment Below 2 lines and comment above 2 lines to run locally
#../lib/aicloudlibs-0.1.0-py3-none-any.whl  
#../lib/adversarial_robustness_toolbox-1.16.0-py3-none-any.whl
